requirements.txt
aenum==3.1.15
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
decorator==5.1.1
dj-database-url==2.1.0
Django==5.0.4
django-dotenv==1.4.2
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.1
exceptiongroup==1.2.2
executing==2.1.0
idna==3.7
ipython==8.27.0
ipython-genutils==0.2.0
jedi==0.19.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==2.1.2
oauthlib==3.2.2
packaging==24.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.1
prompt_toolkit==3.0.47
psycopg==3.1.18
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
xmltodict==0.13.0
